Inital commit
Summary:

This commit includes the initial files for this project. At present, the
program will stream submissions from a subreddit, and if certain
keywords are in the title, then it will log that submission into a CSV
file for later use.

Testing:

Verified that it streams submissions from a subreddit and logs those
that contain specific words in their titles.